Resolved Logic Problems in Test-TargetResource
Test-TargetResource previously had a logic problem when more than three
parameters were specified would cause it to return a false negative.
This was a common occurrence considering that function has 11
parameters. The new logic makes ensure present the default if not
specified. It also compares only the specified parameters and returns
true if they match. It now calls Get-TargetResource and compares the
results to the input parameters to determine if the share is in the
desired state.